[lib] Add chat completion service
