+++ packages/lib/package.json
@@ -69,7 +69,6 @@
         "typescript": "^4.9.5"
     },
     "dependencies": {
-        "gpt-token-utils": "^1.2.0",
         "uuid": "^9.0.0",
         "zod": "^3.21.4"
     },
