+++ packages/lib/src/classes/index.ts
@@ -1,3 +1,8 @@
+export * from "./ConversationConfig.js";
+export * from "./ConversationRequestOptions.js";
+export * from "./ConversationHistory.js";
+export * from "./ConversationCallableFunctions.js";
+export * from "./ChatCompletionService.js";
 export * from "./Conversation.js";
 export * from "./ConversationConfig.js";
 export * from "./Message.js";
