+++ packages/lib/src/utils/index.ts
@@ -1,4 +1,3 @@
-export { default as getMessageTokens } from "./getMessageTokens.js";
 export { default as createChatCompletion } from "./createChatCompletion.js";
 export { default as createDryChatCompletion } from "./createDryChatCompletion.js";
 export { default as getRequestHeaders } from "./getRequestHeaders.js";
